[fix] API patch for email address error
error occured when an event in in other state than normal, for e.g. private, and EWS SOAP API return an error for all events comes next after theses events. Patch consist here to do a while loop, and crop the array at the index of the error to continue the API call for events coming next after a private event or someting like